backend/kudos_project/requirements.txt
Django
djangorestframework
django-cors-headers
python-dateutil